feat: Integrate project name for font export
Pass the project name down to the export actions hook to be used as the base filename for downloaded OTF files. This allows users to export their font files with a filename that reflects the project they are working on.

The "Test Page" tab has also been removed from the settings as it's no longer relevant.