feat: Introduce preferKerningTerm setting
Adds a new `preferKerningTerm` boolean setting to the application. This setting allows users to choose whether to use the term "Kerning" or "Spacing" in the UI, even when in "simple" editor mode.

This change refactors UI elements and logic in `App.tsx`, `AppHeader.tsx`, `KerningPage.tsx`, and `KerningWorkspace.tsx` to conditionally display or label based on the `editorMode` or the new `preferKerningTerm` setting. The `SwitchScriptIcon` component also now accepts a `className` prop for better flexibility.