feat: Add specific mark-base pair exceptions
Introduces the `exceptPairs` property to `AttachmentClass` for defining exceptions to mark attachment rules. This allows for more granular control over positioning by specifying which base characters a mark should not attach to, or vice-versa, within a given attachment class.

Updates the `positioningService` to check for and respect these new pair exceptions during rule application.

Also includes minor UI and styling improvements to the `PositioningToolbar` and fixes a bug in the `ScriptSelection` component related to checkbox event handling.